// lib/api/schemas/stageResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Onboarding-as-a-Service
 * Onboarding-as-a-Service (OaaS) Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { StageResponseDescription } from './stageResponseDescription';
import type { StageResponseType } from './stageResponseType';
import type { StageResponseContentBlocksItem } from './stageResponseContentBlocksItem';

export interface StageResponse {
  name: string;
  description?: StageResponseDescription;
  type?: StageResponseType;
  id: string;
  flow_id: string;
  order: number;
  status: string;
  created_at: string;
  content_blocks?: StageResponseContentBlocksItem[];
}