// lib/api/schemas/contentBlockResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Onboarding-as-a-Service
 * Onboarding-as-a-Service (OaaS) Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { ContentBlockResponseConfig } from './contentBlockResponseConfig';
import type { ContentBlockResponseContent } from './contentBlockResponseContent';

export interface ContentBlockResponse {
  type: string;
  config: ContentBlockResponseConfig;
  content: ContentBlockResponseContent;
  order_index: number;
  id: string;
  stage_id: string;
  created_at: string;
  updated_at: string;
}