// lib/api/schemas/contentTypeResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Onboarding-as-a-Service
 * Onboarding-as-a-Service (OaaS) Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { ContentTypeResponseDescription } from './contentTypeResponseDescription';
import type { ContentTypeResponseCategory } from './contentTypeResponseCategory';
import type { ContentTypeResponseDefaultConfig } from './contentTypeResponseDefaultConfig';

export interface ContentTypeResponse {
  name: string;
  display_name: string;
  description?: ContentTypeResponseDescription;
  category?: ContentTypeResponseCategory;
  default_config?: ContentTypeResponseDefaultConfig;
  id: string;
  is_active: boolean;
  created_at: string;
}